Refactor/Rename package
