+++ pom.xml
@@ -1,3 +1,7 @@
+<!--
+  ~ Copyright (c) 2017. @author lev-gc
+  -->
+
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 
@@ -16,7 +20,6 @@
             <artifactId>junit</artifactId>
             <version>4.3</version>
         </dependency>
-
     </dependencies>
 
     <build>
