+++ src/main/java/own/lesson/learn/common/PhotoUploadUtils.java
@@ -1,0 +1,19 @@
+/*
+ * Copyright (c) 2017. @author lev-gc
+ */
+
+package own.lesson.learn.common;
+
+/**
+ * <b><code>PhotoUploadUtils</code></b>
+ * <p/>
+ * PhotoUploadUtils
+ * <p/>
+ * <b>Creation Time:</b> 2016/5/30 11:09.
+ *
+ * @author lev-gc
+ * @version $Revision$ 2016/5/30 11:09.
+ * @since LessonLearn 1.0
+ */
+public class PhotoUploadUtils {
+}
