--- src/main/java/own/lessonlearn/common/PhotoUploadUtils.java
@@ -1,19 +1,0 @@
-/*
- * Copyright (c) 2016, lev-gc. All Rights Reserved.
- */
-
-package own.lessonlearn.common;
-
-/**
- * <b><code>PhotoUploadUtils</code></b>
- * <p/>
- * PhotoUploadUtils
- * <p/>
- * <b>Creation Time:</b> 2016/5/30 11:09.
- *
- * @author lev-gc
- * @version $Revision$ 2016/5/30 11:09.
- * @since LessonLearn 1.0
- */
-public class PhotoUploadUtils {
-}
